+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file. This
 project adheres to [Semantic Versioning](http://semver.org).
 
+## Unreleased
+
+### Added
+
+  - HTML, script and stylesheet for the item page.
+
 ## 0.4.0 - 2016-12-06
 
 ### Added